Added a basic malloc, updated build script, various small improvements.

+++ compile.sh
@@ -9,7 +9,7 @@ trap finish EXIT
 set -e
 
 echo 'starting compilation'
-rm os-image || true
+rm -f os-image || true
 
 echo 'running assembler'
 nasm kernel_entry.asm -f elf -o kernel_entry.o
@@ -23,7 +23,15 @@ ld -melf_i386 -r -b binary -o KFS.o KFS.bin
 echo "compiling kernel"
 gcc -ffreestanding -m32 -c kernel.c -o kernel.o
 ld -melf_i386 -o kernel.bin -Ttext 0x8000 kernel_entry.o int.o kernel.o KFS.o --oformat binary --entry main
-dd if=kernel.bin count=100 of=padded_kernel.bin conv=sync &> /dev/null && sync
+
+sectorcount=100
+size=$(wc -c < kernel.bin)
+if ((size<$(($sectorcount*512)))) ; then
+	dd if=kernel.bin count=$sectorcount of=padded_kernel.bin conv=sync &> /dev/null && sync
+else
+	echo "kernel image is truncated, you must increase the padding count!"
+	exit 1
+fi
 
 echo "assembling bootloader"
 if [ "$use_new_bootloader" = true ]; then
@@ -60,10 +68,25 @@ dd if=/dev/zero of=floppy.img bs=512 count=2880 &> /dev/null && sync
 dd if=os-image of=floppy.img conv=notrunc &> /dev/null && sync
 if [ "$1" == "write" ]
 then
-	sudo dd if=floppy.img of=/dev/sdb conv=notrunc && sync
+	echo "You will overwrite /dev/sdb: $(ls -l /dev/disk/by-id | grep sdb | cut -d ' ' -f 11) "
+	read -p "Are you sure? (y/n)" -n 1 -r
+	echo 
+	if [[ $REPLY =~ ^[Yy]$ ]]
+	then
+		sudo dd if=floppy.img of=/dev/sdb conv=notrunc && sync
+	else
+		buildstatus="canceled"
+		exit 2
+	fi
 else
 	qemu-system-i386 floppy.img -soundhw pcspk &
 fi
+
+# increment buildData.c values
+cp buildData.c buildData-temp.c
+awk '/int buildID = [0-9]+;/ { printf "int buildID = %d;\n", $4+1 };!/int buildID = [0-9]+;/{print}' < buildData-temp.c > buildData.c
+rm buildData-temp.c
+
 buildstatus="successful"
 rm *.o
 rm *.bin
